requirements.txt
backports.functools-lru-cache==1.5
boto3==1.9.200
botocore==1.12.200
configparser==3.7.4
Django==1.11.16
django-extensions==2.2.1
django-storages==1.7.1
djangorestframework==3.9.4
docutils==0.14
entrypoints==0.3
enum34==1.1.6
flake8==3.7.8
functools32==3.2.3.post2
futures==3.3.0
isort==4.3.21
jmespath==0.9.4
mccabe==0.6.1
psycopg2==2.8.3
pycodestyle==2.5.0
pyflakes==2.1.1
python-dateutil==2.8.0
pytz==2019.2
s3transfer==0.2.1
six==1.12.0
typing==3.7.4
urllib3==1.25.3
